APIs/API_ImageRecognitionAI/requirements.txt
google-genai
Pillow
requests
pandas
werkzeug
